Re VIV-481: Store folding state globally, not per article
(Yes, this is the opposite of RED-1944 and commit:0f54374cad77,
the requirement / user requested behaviour simply has changed in the meantime.)

Note: window.sessionStorage is scoped to the browser window, so we use localStorage now.
This has the disadvantage that currently there is no UI to clear the saved state (and
restarting the browser won't do it).